Release 2.0
- Add /introduce command.
- Add /jira command and its alias /bug.
- Add /tic_tac_toe command.
- Add code block message menu.
- Add auto-threading.
- Add question forum management.
- Add message package for message events.
- Rewrite some parts of command block system, add CommandBlocksHandle.LotteryData.
- Use serialize to replace JSON.
- Optimization.

+++ src/main/java/cartoland/commands/OneATwoBCommand.java
@@ -22,7 +22,7 @@ public class OneATwoBCommand implements ICommand
 	private final CommandUsage commandCore;
 	private static final int MAX_MINUTE = 2;
 	private static final int MAX_GUESSES = 7;
-	private static final int REWARD = 100;
+	private static final byte REWARD = 100;
 
 	public OneATwoBCommand(CommandUsage commandUsage)
 	{
